github-app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.5.0
PyJWT>=2.8.0
cryptography>=42.0.0
